refactor: export rpc errors from module instead of submodule
