+++ crates/rust-client/src/rpc/domain/accounts.rs
@@ -2,11 +2,11 @@ use std::fmt::{Debug, Display, Formatter};
 
 use miden_objects::accounts::AccountId;
 
-use crate::rpc::errors::RpcConversionError;
 #[cfg(feature = "tonic")]
 use crate::rpc::tonic_client::generated::account::AccountId as ProtoAccountId;
 #[cfg(feature = "web-tonic")]
 use crate::rpc::web_tonic_client::generated::account::AccountId as ProtoAccountId;
+use crate::rpc::RpcConversionError;
 
 // ACCOUNT ID
 // ================================================================================================
