+++ crates/rust-client/src/rpc/domain/blocks.rs
@@ -1,11 +1,11 @@
 use miden_objects::BlockHeader;
 
 use super::MissingFieldHelper;
-use crate::rpc::errors::RpcConversionError;
 #[cfg(feature = "tonic")]
 use crate::rpc::tonic_client::generated::block_header;
 #[cfg(feature = "web-tonic")]
 use crate::rpc::web_tonic_client::generated::block_header;
+use crate::rpc::RpcConversionError;
 
 // BLOCK HEADER
 // ================================================================================================
