+++ crates/rust-client/src/rpc/domain/digest.rs
@@ -3,11 +3,11 @@ use std::fmt::{Debug, Display, Formatter};
 use hex::ToHex;
 use miden_objects::{notes::NoteId, Digest, Felt, StarkField};
 
-use crate::rpc::errors::RpcConversionError;
 #[cfg(feature = "tonic")]
 use crate::rpc::tonic_client::generated::digest;
 #[cfg(feature = "web-tonic")]
 use crate::rpc::web_tonic_client::generated::digest;
+use crate::rpc::RpcConversionError;
 
 // CONSTANTS
 // ================================================================================================
