+++ crates/rust-client/src/rpc/domain/merkle.rs
@@ -3,11 +3,11 @@ use miden_objects::{
     Digest,
 };
 
-use crate::rpc::errors::RpcConversionError;
 #[cfg(feature = "tonic")]
 use crate::rpc::tonic_client::generated;
 #[cfg(feature = "web-tonic")]
 use crate::rpc::web_tonic_client::generated;
+use crate::rpc::RpcConversionError;
 
 // MERKLE PATH
 // ================================================================================================
