+++ crates/rust-client/src/rpc/domain/nullifiers.rs
@@ -1,10 +1,10 @@
 use miden_objects::{crypto::hash::rpo::RpoDigest, notes::Nullifier};
 
-use crate::rpc::errors::RpcConversionError;
 #[cfg(feature = "tonic")]
 use crate::rpc::tonic_client::generated::digest::Digest;
 #[cfg(feature = "web-tonic")]
 use crate::rpc::web_tonic_client::generated::digest::Digest;
+use crate::rpc::RpcConversionError;
 
 // INTO NULLIFIER
 // ================================================================================================
