+++ crates/rust-client/src/rpc/domain/transactions.rs
@@ -1,6 +1,5 @@
 use miden_objects::{crypto::hash::rpo::RpoDigest, transaction::TransactionId};
 
-use crate::rpc::errors::RpcConversionError;
 #[cfg(feature = "tonic")]
 use crate::rpc::tonic_client::generated::{
     digest::Digest, transaction::TransactionId as ProtoTransactionId,
@@ -9,6 +8,7 @@ use crate::rpc::tonic_client::generated::{
 use crate::rpc::web_tonic_client::generated::{
     digest::Digest, transaction::TransactionId as ProtoTransactionId,
 };
+use crate::rpc::RpcConversionError;
 
 // INTO TRANSACTION ID
 // ================================================================================================
